Replace webpack dev server with our own
That we also use as a target for our nightwatch
tests.
This is a good candidate to move upstream to the 
maji lib project
